Paratransit options now have a configurable minimum duration, which defaults to 2 hours (in _oneclick.rb).  So they now have starts & ends.
In addition, unless SKIP_DYNAMIC_PARATRANSIT_DURATION is set, we also
hit OTP to get car trip time, and multiply that by a factor (also
configurable, system-wide at the moment) to get a dynamic estimate.

At the moment with the factor set to 2x, in PA at least most trips
are using the static minimum duration.

[Fixes #71222998]